refactor(temporal): rename and extract in mtv_routes.go every mtv related route and handler
